requirements.txt
streamlit
pydicom
wfdb
numpy
pillow
matplotlib
scipy
google-generativeai
